requirements.txt
pandas>=2.2.0
pyyaml>=6.0.1
sqlfluff>=3.0.0
jinja2>=3.1.2
click>=8.1.7
python-dotenv>=1.0.0
pydantic>=2.4.2
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.3.0
pycurl>=7.45.2 